// apps/hypersync-indexer/generated/src/db/Enums.gen.ts
/* TypeScript file generated from Enums.res by genType. */

/* eslint-disable */
/* tslint:disable */

export type ContractType_t = "ENSGovernor" | "ENSToken";

export type EntityType_t = 
    "Account"
  | "AccountBalance"
  | "AccountPower"
  | "BalanceHistory"
  | "DaoMetricsDayBucket"
  | "Delegation"
  | "FeedEvent"
  | "ProposalOnchain"
  | "Token"
  | "TokenPrice"
  | "Transaction"
  | "Transfer"
  | "VoteOnchain"
  | "VotingPowerHistory"
  | "dynamic_contract_registry";

export type EventType_t = 
    "VOTE"
  | "PROPOSAL"
  | "PROPOSAL_EXTENDED"
  | "DELEGATION"
  | "DELEGATION_VOTES_CHANGED"
  | "TRANSFER";

export type MetricType_t = 
    "total"
  | "delegated"
  | "cex"
  | "dex"
  | "lending"
  | "circulating"
  | "treasury"
  | "non_circulating";